farzone/requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.3
daphne==4.2.0
decorator==5.2.1
Django==5.1.7
django-appconf==1.1.0
django-extensions==4.1
django-image-cropping==1.7
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
easy-thumbnails==2.10
executing==2.2.0
fastapi==0.115.0
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
msgpack==1.1.0
multidict==6.4.4
parso==0.8.4
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.1
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.1.0
python-dotenv==1.0.1
python-telegram-bot==22.1
redis==6.1.0
requests==2.32.3
service-identity==24.2.0
setuptools==80.8.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.38.6
telegram==0.0.1
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.31.0
wcwidth==0.2.13
yarl==1.20.0
zope.interface==7.2
